Replace SplashScreen with Android12 version
